refactor: give inference ownership of networking and build inputs
