+++ inference/nix/patches/nvattest-ca-bundle.patch
