+++ inference/nix/patches/nvattest-explicit-runtime-links.patch
