+++ inference/nix/patches/nvattest-pinned-sources.patch
