+++ inference/nix/patches/regorus-pinned-revision.patch
