+++ inference/rootfs/etc/nftables.conf
