+++ sandbox/cmd/shim/main.go
@@ -4,7 +4,6 @@ import (
 	"strconv"
 
 	"tinfoil/internal/attestation"
-	"tinfoil/internal/containernet"
 	"tinfoil/shim"
 )
 
@@ -15,7 +14,7 @@ var sandboxPorts = []int{22, 3000, 3001, 4000, 5000, 5173, 8000, 8001, 8888, 900
 
 func shimSpec() shim.Spec {
 	return shim.Spec{
-		UpstreamHost:   upstreamHost,
+		UpstreamHost:   "127.0.0.1",
 		PublishedPorts: publishedPorts,
 		Observability: shim.Observability{
 			DeviceEvidence: attestation.NoDeviceEvidence,
@@ -25,9 +24,7 @@ func shimSpec() shim.Spec {
 
 func main() { shim.Main(shimSpec()) }
 
-func upstreamHost(string) string { return containernet.PublishedHostIP }
-
-func publishedPorts(string) (map[string]bool, error) {
+func publishedPorts() (map[string]bool, error) {
 	targets := make(map[string]bool, len(sandboxPorts))
 	for _, port := range sandboxPorts {
 		targets[strconv.Itoa(port)] = true
