+++ sandbox/rootfs/etc/nftables.conf
@@ -1,0 +1,51 @@
+#!/usr/sbin/nft -f
+
+flush ruleset
+
+table inet tinfoil {
+    # HTTP-01 is opened temporarily by tinfoil-boot and then flushed closed.
+    chain http01 {
+    }
+
+    chain inbound {
+    }
+
+    chain container_input {
+    }
+
+    chain container_forward {
+    }
+
+    chain input {
+        type filter hook input priority 0; policy drop;
+
+        # Loopback (shim ↔ upstream containers)
+        iif "lo" accept
+
+        # Established/related (return traffic for outbound connections)
+        ct state established,related accept
+
+        jump http01
+
+        jump inbound
+
+        jump container_input
+
+        # ICMP/ICMPv6 (network diagnostics)
+        ip protocol 1 accept
+        meta l4proto 58 accept
+
+        # Shim TLS port (always open)
+        tcp dport 443 accept
+    }
+
+    chain forward {
+        type filter hook forward priority 0; policy drop;
+
+        jump container_forward
+    }
+
+    chain output {
+        type filter hook output priority 0; policy accept;
+    }
+}
