+++ tinfoil/boot/cert_test.go
@@ -3,8 +3,6 @@ package boot
 import (
 	"crypto/tls"
 	"errors"
-	"os"
-	"path/filepath"
 	"reflect"
 	"strings"
 	"testing"
@@ -102,33 +100,5 @@ func TestWithHTTP01FirewallPreservesRequestAndCleanupFailures(t *testing.T) {
 	}
 	if !errors.Is(err, cleanupErr) {
 		t.Fatalf("error = %v, want wrapped %v", err, cleanupErr)
-	}
-}
-
-func TestMeasuredFirewallDefinesHTTP01Chain(t *testing.T) {
-	policyPath := filepath.Join("..", "..", "image", "rootfs", "etc", "nftables.conf")
-	policy, err := os.ReadFile(policyPath)
-	if err != nil {
-		t.Fatalf("reading %s: %v", policyPath, err)
-	}
-
-	text := string(policy)
-	for _, contract := range []string{
-		"chain http01 {\n    }",
-		"jump http01",
-		"ip protocol 1 accept",
-		"meta l4proto 58 accept",
-	} {
-		if !strings.Contains(text, contract) {
-			t.Fatalf("measured firewall policy does not contain %q", contract)
-		}
-	}
-	if strings.Contains(text, "tcp dport 80 accept") {
-		t.Fatal("measured firewall policy opens HTTP-01 before certificate issuance")
-	}
-	for _, protocolName := range []string{"ip protocol icmp", "l4proto ipv6-icmp"} {
-		if strings.Contains(text, protocolName) {
-			t.Fatalf("measured firewall policy depends on protocol-name parsing: %q", protocolName)
-		}
 	}
 }
